导航/requirements.txt
Flask==1.1.4
Werkzeug==1.0.1
Jinja2==2.11.3
MarkupSafe==1.1.1
itsdangerous==1.1.0
requests==2.27.1
python-dotenv==0.19.2
Flask-Cors==3.0.10